test(LottoMachineTest): LottoMachineTest WinningLottoTest와 분리
